Switched to BigDecimal instead of Double.

+++ src/me/greatman/Craftconomy/BalanceCollection.java
@@ -1,13 +1,15 @@
 package me.greatman.Craftconomy;
 
+import java.math.BigDecimal;
+
 public class BalanceCollection
 {
 
 	private String world;
-	private double balance;
+	private BigDecimal balance;
 	private Currency currency;
 
-	public BalanceCollection(String worldName, double balanceAmount, Currency currency)
+	public BalanceCollection(String worldName, BigDecimal balanceAmount, Currency currency)
 	{
 		this.world = worldName;
 		balance = balanceAmount;
@@ -19,7 +21,7 @@ public class BalanceCollection
 		return Craftconomy.plugin.getServer().getWorld(world).getName();
 	}
 
-	public double getBalance()
+	public BigDecimal getBalance()
 	{
 		return balance;
 	}
